refactor: create custom hooks to handle tooltip position and visibility
